# Interactive tracks performance improvements