# Description

This PR proposes a few minor changes that improve the efficiency of `set_data` for the `InteractiveTrackManager`.

I ran the following benchmark command: `asv run --python=same -b time_interactive_manager`

Before changes proposed here:

```
[100.00%] ··· benchmark_tracks_manager.TrackManagerSuite.time_interactive_manager                                                                                                                                              [100.00%] ··· ========= ============= ============= ============= ============
              --                          sorted / n_tracks                   
              --------- ------------------------------------------------------
                 size     False / 10   False / 100    True / 10    True / 100 
              ========= ============= ============= ============= ============
                 100     2.43±0.08ms     11.4±2ms    2.39±0.05ms   9.76±0.3ms 
                 1000     5.40±0.1ms    18.7±0.2ms   5.42±0.06ms   18.7±0.3ms 
                10000     34.2±0.2ms    49.2±0.4ms    34.2±0.3ms   48.8±0.4ms 
                100000     354±3ms       355±2ms      346±0.7ms     345±3ms   
               1000000    3.73±0.03s    3.71±0.01s    3.46±0.01s    3.48±0s   
              ========= ============= ============= ============= ============
```


After changes proposed here all except a few of the size=100 cases are faster:
```
[100.00%] ··· benchmark_tracks_manager.TrackManagerSuite.time_interactive_manager                                                                                                                                              [100.00%] ··· ========= ============= ============= ============= =============
              --                           sorted / n_tracks                   
              --------- -------------------------------------------------------
                 size     False / 10   False / 100    True / 10     True / 100 
              ========= ============= ============= ============= =============
                 100     2.58±0.02ms    9.91±0.3ms   2.59±0.03ms    9.94±0.5ms 
                 1000    5.13±0.05ms   17.1±0.09ms    5.15±0.1ms    17.9±0.5ms 
                10000     31.2±0.8ms    39.4±0.5ms    29.5±0.3ms   38.9±0.07ms 
                100000     319±2ms       275±4ms       295±4ms       259±3ms   
               1000000    3.51±0.02s    2.91±0.01s    3.01±0.03s    2.53±0.01s 
              ========= ============= ============= ============= =============
```

I am not sure we want the change in d49fd51, though. If I revert that change you will see that performance is bit worse in some cases, but better in others. Result without d49fd51:

```
[100.00%] ··· benchmark_tracks_manager.TrackManagerSuite.time_interactive_manager
[100.00%] ··· ========= ============= ============= ============= ============
              --                          sorted / n_tracks                   
              --------- ------------------------------------------------------
                 size     False / 10   False / 100    True / 10    True / 100 
              ========= ============= ============= ============= ============
                 100     2.30±0.02ms    10.7±0.7ms   2.45±0.08ms   9.71±0.1ms 
                 1000     4.80±0.1ms    17.5±0.2ms   4.46±0.03ms   17.7±0.2ms 
                10000     26.5±0.2ms    41.6±0.3ms    25.5±0.3ms   40.4±0.2ms 
                100000     274±3ms       278±1ms       267±2ms      263±3ms   
               1000000    2.91±0.02s    2.93±0.02s     2.63±0s      2.71±0s   
              ========= ============= ============= ============= ============
```