@JoOkuma 2022-01-10
Hi Gregory, thank you for this PR.

I think we could undo d49fd51, the benefits are very small for the 100 tracks.

The rest is very good. I didn't know named arguments added some overhead, I won't be using them that often now :P