@grlee77 2022-01-10
> I didn't know named arguments added some overhead, I won't be using them that often now :P

It is fairly minimal overhead that I would typically ignore unless you are calling it millions of times (possible here since it is called for every node). Even then I only measured around 0.1 second improvement at 1,000,000 nodes with that change. It is the first two commits here that provide most of the benefit.