@grlee77 2022-01-10
I have reverted the change to the _time_to_nodes computation